requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.27.0
mcp==1.1.2
pydantic==2.10.4
python-dotenv==1.0.1
anthropic==0.42.0
ollama==0.4.4
qrcode==8.0
